% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram.R
\name{get_ngram}
\alias{get_ngram}
\title{Hidden internal function to get NGRAM}
\usage{
get_ngram(word, corpus, language = NULL)
}
\arguments{
\item{word}{The word to get NGRAM for.}

\item{corpus}{The corpus to use. Options are 'avis' and 'bok'.}

\item{language}{The language to use. Default is NULL.}
}
\value{
A list that contains the NGRAM.
}
\description{
Hidden internal function to get NGRAM
}
\keyword{internal}
