% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram.R
\name{ngram_conv}
\alias{ngram_conv}
\title{Hidden internal function to convert NGRAM}
\usage{
ngram_conv(ngrams, smooth = 1, years = c(1810, 2013), mode = "relative")
}
\arguments{
\item{ngrams}{The NGRAMs to convert.}

\item{smooth}{Smoothing factor. Default is 1.}

\item{years}{A vector that contains the start and end years. Default is c(1810,2013).}

\item{mode}{The mode to use. Default is 'relative'.}
}
\value{
A data frame that contains the converted NGRAM.
}
\description{
Hidden internal function to convert NGRAM
}
\keyword{internal}
